refactor(messages): move WhatsApp message handlers to dedicated folder\n\n- Moved sendMessage, replyMessage, sendList, sendReplyButtons to src/messages/messageHandlers.js\n- Updated app.js to import from new location\n\nCloses #message-refactor
