+++ src/messages/messageHandlers.js
@@ -1,0 +1,156 @@
+const axios = require('axios');
+
+const WHATSAPP_ACCESS_TOKEN = process.env.WHATSAPP_TOKEN;
+
+async function sendMessage(to, body) {
+  await axios({
+    url: `https://graph.facebook.com/v21.0/${process.env.WHATSAPP_PHONE_NUMBER_ID}/messages`,
+    method: 'post',
+    headers: {
+      'Authorization': `Bearer ${WHATSAPP_ACCESS_TOKEN}`,
+      'Content-Type': 'application/json'
+    },
+    data: JSON.stringify({
+      messaging_product: 'whatsapp',
+      to,
+      type: 'text',
+      text: {
+        body
+      }
+    })
+  })
+}
+
+async function replyMessage(to, body, messageId) {
+  await axios({
+    url: `https://graph.facebook.com/v21.0/${process.env.WHATSAPP_PHONE_NUMBER_ID}/messages`,
+    method: 'post',
+    headers: {
+      'Authorization': `Bearer ${WHATSAPP_ACCESS_TOKEN}`,
+      'Content-Type': 'application/json'
+    },
+    data: JSON.stringify({
+      messaging_product: 'whatsapp',
+      to,
+      type: 'text',
+      text: {
+        body
+      },
+      context: {
+        message_id: messageId
+      }
+    })
+  })
+}
+
+async function sendList(to) {
+  await axios({
+    url: `https://graph.facebook.com/v21.0/${process.env.WHATSAPP_PHONE_NUMBER_ID}/messages`,
+    method: 'post',
+    headers: {
+      'Authorization': `Bearer ${WHATSAPP_ACCESS_TOKEN}`,
+      'Content-Type': 'application/json'
+    },
+    data: JSON.stringify({
+      messaging_product: 'whatsapp',
+      to,
+      type: 'interactive',
+      interactive: {
+        type: 'list',
+        header: {
+          type: 'text',
+          text: 'Message Header'
+        },
+        body: {
+          text: 'This is a interactive list message'
+        },
+        footer: {
+          text: 'This is the message footer'
+        },
+        action: {
+          button: 'Tap for the options',
+          sections: [
+            {
+              title: 'First Section',
+              rows: [
+                {
+                  id: 'first_option',
+                  title: 'First option',
+                  description: 'This is the description of the first option'
+                },
+                {
+                  id: 'second_option',
+                  title: 'Second option',
+                  description: 'This is the description of the second option'
+                }
+              ]
+            },
+            {
+              title: 'Second Section',
+              rows: [
+                {
+                  id: 'third_option',
+                  title: 'Third option'
+                }
+              ]
+            }
+          ]
+        }
+      }
+    })
+  })
+}
+
+async function sendReplyButtons(to) {
+  await axios({
+    url: `https://graph.facebook.com/v21.0/${process.env.WHATSAPP_PHONE_NUMBER_ID}/messages`,
+    method: 'post',
+    headers: {
+      'Authorization': `Bearer ${WHATSAPP_ACCESS_TOKEN}`,
+      'Content-Type': 'application/json'
+    },
+    data: JSON.stringify({
+      messaging_product: 'whatsapp',
+      to,
+      type: 'interactive',
+      interactive: {
+        type: 'button',
+        header: {
+          type: 'text',
+          text: 'Message Header'
+        },
+        body: {
+          text: 'This is a interactive reply buttons message'
+        },
+        footer: {
+          text: 'This is the message footer'
+        },
+        action: {
+          buttons: [
+            {
+              type: 'reply',
+              reply: {
+                id: 'first_button',
+                title: 'First Button'
+              }
+            },
+            {
+              type: 'reply',
+              reply: {
+                id: 'second_button',
+                title: 'Second Button'
+              }
+            }
+          ]
+        }
+      }
+    })
+  })
+}
+
+module.exports = {
+  sendMessage,
+  replyMessage,
+  sendList,
+  sendReplyButtons
+};
